refactor: colocate web formatting utils
