+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/settings/organization/-members/invite-members-form.tsx
@@ -14,7 +14,7 @@ import {
 import { Field, FieldError, FieldLabel } from "@packages/ui/components/field";
 import { Spinner } from "@packages/ui/components/spinner";
 import { Textarea } from "@packages/ui/components/textarea";
-import { getInitials } from "@/lib/initials";
+import { getInitials } from "@/utils/text";
 import { useForm } from "@tanstack/react-form";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { toast } from "@packages/ui/hooks/use-toast";
