+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/settings/profile.tsx
@@ -23,7 +23,7 @@ import { Label } from "@packages/ui/components/label";
 import { PasswordInput } from "@packages/ui/components/password-input";
 import { Separator } from "@packages/ui/components/separator";
 import { Skeleton } from "@packages/ui/components/skeleton";
-import { getInitials } from "@/lib/initials";
+import { getInitials } from "@/utils/text";
 import { useForm } from "@tanstack/react-form";
 import { useMutation, useSuspenseQuery } from "@tanstack/react-query";
 import { createFileRoute } from "@tanstack/react-router";
