+++ apps/web/src/utils/date.ts
@@ -1,0 +1,5 @@
+import dayjs from "dayjs";
+
+export function formatDate(date: Date | string) {
+   return dayjs(date).format("DD/MM/YYYY");
+}
